Split root test_deployment.py
